transaction: Move Error type to mod error
Mirror the style used by other subcommand modules.
