+++ src/transaction/mod.rs
@@ -1,2 +1,3 @@
 pub mod core;
 pub mod commands;
+pub mod error;
